requirements.txt
# LUINT dependencies
beautifulsoup4>=4.13.0
click>=8.0.0
colorama>=0.4.4
dnspython>=2.0.0
email-validator>=2.0.0
flask>=3.0.0
flask-sqlalchemy>=3.0.0
ipwhois>=1.2.0
jinja2>=3.0.0
mmh3>=3.0.0
psycopg2-binary>=2.9.0
pyopenssl>=21.0.0
python-nmap>=0.7.1
python-whois>=0.7.3
pyyaml>=6.0.0
requests>=2.25.0
requests-cache>=0.9.0
rich>=10.0.0
tldextract>=3.1.0
trafilatura>=2.0.0
urllib3>=1.26.0
tabulate>=0.8.9
cryptography>=35.0.0